Patch interface between BioReactor and Matcher

+++ src/biocatalyzer/bioreactor.py
@@ -79,7 +79,6 @@ class BioReactor:
             self._n_jobs = n_jobs
         self._new_compounds_path = os.path.join(self._output_path, 'new_compounds.tsv')
         self._new_compounds = None
-        self._new_compounds_flag = False
 
     @property
     def compounds(self):
@@ -147,7 +146,7 @@ class BioReactor:
         pd.DataFrame
             The new compounds generated.
         """
-        if self._new_compounds_flag:
+        if self._new_compounds is not None:
             return Loaders.load_compounds(self._new_compounds_path, False)
         else:
             raise ValueError('No compounds generated yet. Run the BioReactor react method first.')
@@ -526,7 +525,7 @@ class BioReactor:
         """
         return self._reaction_rules[self._reaction_rules.InternalID == reaction_rule_id].EC_Numbers.values[0]
 
-    def process_results(self, save: bool = True):
+    def process_results(self, save: bool = True, overwrite: bool = True):
         """
         Process the results of the reactor.
         Group results by unique SMILES and merges the other columns.
@@ -535,11 +534,13 @@ class BioReactor:
         ----------
         save: bool
             If True, save the results to a file.
+        overwrite: bool
+            If True, overwrite the results file if it already exists.
 
         Returns
         -------
-        pd.DataFrame
-            The processed results.
+        Tuple[pd.DataFrame, str]
+            The processed results and the path to the results file.
         """
         results = pd.read_csv(self._new_compounds_path, sep='\t', header=0)
         results.EC_Numbers = results.EC_Numbers.fillna('')
@@ -557,9 +558,15 @@ class BioReactor:
         results['NewReactionSmiles'] = results['NewReactionSmiles'].apply(lambda x: _merge_fields(x))
         results['EC_Numbers'] = results['EC_Numbers'].apply(lambda x: _merge_fields(x))
         if save:
-            results_file_proc = os.path.join(self._output_path, 'new_compounds_processed.tsv')
-            results.to_csv(results_file_proc, sep='\t', index=False)
-        return results
+            if overwrite:
+                results_file_proc = os.path.join(self._output_path, 'new_compounds.tsv')
+                results.to_csv(results_file_proc, sep='\t', index=False)
+            else:
+                results_file_proc = os.path.join(self._output_path, 'new_compounds_processed.tsv')
+                results.to_csv(results_file_proc, sep='\t', index=False)
+        else:
+            results_file_proc = self._new_compounds_path
+        return results, results_file_proc
 
     def _react_single(self, smiles: str, smarts: str):
         """
@@ -594,7 +601,6 @@ class BioReactor:
                         with open(self._new_compounds_path, 'a') as f:
                             f.write(f"{smiles_id}\t{smiles}\t{smarts_id}\t{smiles_id}_{uuid.uuid4()}\t"
                                     f"{most_similar_product}\t{result}\t{ecs}\n")
-                        self._new_compounds_flag = True
 
     def react(self):
         """
@@ -610,9 +616,6 @@ class BioReactor:
         self._new_compounds = f"New products saved to {self._new_compounds_path}"
         t1 = time.time()
         logging.info(f"Time elapsed: {t1 - t0} seconds")
-        if self._new_compounds_flag:
-            return True
-        return False
 
 
 if __name__ == '__main__':
